+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+### Added
+
+- Snapshot export/import UI components and wiring in the repository page.
+
 ## [1.1.0] - 2026-01-13
 
 ### Added